requirements.txt
requests>=2.8.1
lxml>=3.6.4
mock>=2.0.0
future>=0.16.0
six>=1.10.0
xmlunittest>=0.3.2
rdflib-jsonld>=0.4.0